crud users OK

+++ weroad-backend/src/users/users.controller.ts
@@ -2,19 +2,23 @@ import {
   Body,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   HttpCode,
+  HttpException,
   HttpStatus,
   Param,
   Post,
   Put,
+  Req,
   UseGuards,
 } from '@nestjs/common';
 import { AuthGuard } from 'src/auth/auth.guard';
 import { AdminGuard } from 'src/roles/roles.guard';
-import { UserDto } from './dto/user.dto';
+import { UserDto, UserUpdateDto } from './dto/user.dto';
 import { UsersService } from './users.service';
 import { User } from './entities/user.entity';
+import { Request } from 'express';
 
 @Controller('admin/users')
 @UseGuards(AuthGuard, AdminGuard)
@@ -37,13 +41,22 @@ export class UsersController {
   }
 
   @Put(':id')
-  update(@Param('id') id: string, @Body() userDto: UserDto): Promise<User> {
+  update(
+    @Param('id') id: string,
+    @Body() userDto: UserUpdateDto,
+  ): Promise<User> {
     return this.usersService.update(id, userDto);
   }
 
   @HttpCode(HttpStatus.OK)
   @Delete(':id')
-  remove(@Param('id') id: string): Promise<void> {
+  remove(@Param('id') id: string, @Req() request: Request): Promise<void> {
+    if (request['user'].id === id) {
+      throw new HttpException(
+        'You cannot remove yourself',
+        HttpStatus.FORBIDDEN,
+      );
+    }
     return this.usersService.remove(id);
   }
 }
